experiment-analysis-stack/requirements.txt
boto3
botocore
kafka-python
urllib3